feat(wolf): 实现狼的 interactMob() 驯服交互逻辑
完整实现 WolfEntity::interactMob()，包含以下交互功能：

1. 未驯服 + 骨头 → 1/3 概率驯服，成功时设置主人、广播心形粒子、
   触发 TameAnimalTrigger 进度、默认坐下；失败时广播烟雾粒子
2. 已驯服 + 食物 + 未满血 → 喂食治疗（2.0 * foodNutrition）
3. 已驯服 + 染料 + 主人 → 改变颈圈颜色
4. 已驯服 + 繁殖物品 + 满血 + 成年 → 进入求爱状态
5. 已驯服 + 繁殖物品 + 幼年 → 加速成长
6. 已驯服 + 主人 + 非特殊物品 → 切换坐下/站起

其他改进：
- 将颈圈颜色类型从 u8 改为 DyeColor 枚举，增强类型安全
- 添加 _tryToTame() 私有方法封装驯服概率判定逻辑
- 添加 _getFoodHealAmount() 硬编码狼食物治疗量映射
- 添加 _getDyeColorFromItem() 染料物品到颜色的静态映射
- 注册 ResetAngerGoal<WolfEntity> 到目标选择器优先级8
- 在 TargetGoals.cpp 中添加 WolfEntity 的显式模板实例化
- 更新 WolfCollarLayer 适配 DyeColor 类型
- 更新 WolfEntityTest 适配 DyeColor 类型

+++ src/client/renderer/trident/entity/layer/entity/WolfCollarLayer.cpp
@@ -140,7 +140,7 @@ bool WolfCollarLayer::shouldRender(const ::mc::WolfEntity& entity) const
 
 Vector3f WolfCollarLayer::_getCollarColor(const ::mc::WolfEntity& entity)
 {
-    u8 colorIndex = entity.getCollarColor();
+    u8 colorIndex = static_cast<u8>(entity.getCollarColor());
     if (colorIndex < 16) {
         return COLLAR_COLORS[colorIndex];
     }
